requirements.txt
clams-python==1.0.2
inaSpeechSegmenter==0.7.6
tensorflow==2.*
